Fix postgres unbind for legacy bindings.
We still have postgres instances with bindings that were created before
we introduced the event triggers based permissions model. These bindings
are therefore using the shared rdsbroker_UUID_owner user.

Unbinding is currently failing for these users because it's trying to
drop the username derived from the binding ID, which doesn't exist. We
fix this by partially reverting 05942c88 so that DropUser doesn't error
if the role doesn't exist. (It's not safe to drop the rdsbroker_UUID_owner
user because it could be shared by multiple bindings)