Expand DBInstanceStatus to LastOperation mappings
Previous we only handled translating a subset of the documented
DBInstanceStatus values to LastOperationState responses and considered
all other values to be failure cases.

The list of possible DBInstanceStatus[1] appears to be getting extended
over time and not all of the values should be considered "failures".

We had an issue where a status of "storage-optimization" was causing an
upgraded service to present with a "failed" state.

To protect against this we have updated the list status mappings to all
the currently known values and we will now consider any unknown status
to be "InProgress" responses which hold less risk and ensure polling
will continue.

[1] https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/Overview.DBInstance.Status.html