Implement a cron process for deleting old RDS snapshots
We add a "-cron" flag to the main executable to start a cron process which will regularly delete old RDS snapshots.

We need to delete the old snapshots so we don't reach the AWS limits and also to comply with GDPR requirements.

The following config parameters are needed:
 - cron_schedule: crontab-like expression with seconds precision (e.g. "0 0 * * * *" or "@hourly"), with fields: "second minute hour dom month dow", see [1]
 - keep_snapshots_for_days: only snapshots older than this value will be deleted

We will only delete RDS snapshot which has the "Broker Name" tag matching the broker's name (config: rds_config.broker_name).

[1] https://github.com/robfig/cron/blob/2315d5715e36303a941d907f038da7f7c44c773b/parser.go#L184