Migrate from godep to dep
The godep project is not going to be maintained anymore as the project is going
to be archived. [1]

Also I had issues to add a new dependency as godep was complaining about a
non-existing aws sdk subpackage.

We will migrate to dep as it's the most widely supported package manager and we
use it in almost all of our projects.

[1] https://github.com/tools/godep#godep---archived