Mysql engine return LoginFailedError
Mysql engine would check if the error connecting was due wrong
credentials and return the error LoginFailedError. This error
would trigger a reset master password when the broker starts,
allowing change the seed or master password generation algorithm.

We add github.com/VividCortex/mysqlerr that maintains a list of
the mysql error codes.