Use postgres with auth enabled in unit tests
In order to test the wrong master password credentials, we must
enable the authentication in postgresql. Otherwise, any credentials
would be accepted.

We update the README to start docker with creds, and add some
logic in travis to enable auth by updating the pg_hba.conf and
setting a password for the postgres user.

We update the test in postgres for the tests that check if the
users were deleted, as they would now raise an authentication error
when testing with a auth enabled postgres, instead of a role not found.