Integration test for MasterPasswordSeed reset
We want to be sure that the MasterPassword is reset if
the value of the MasterPasswordSeed is updated, and that
the broker can still bind and unbind after the change.

To support this, we must create a new broker daemon process
per test, so that we can restart it to update the MasterPassword
seed. We move the logic to start/stop the broker to the tests
as a Before/After each.

The operation to update the MasterPassword runs asynchronously,
so any operation would fail for a while. We use Eventually to
retry for 2 minutes the unbind operation after rotate the
MasterPasswordSeed.

Existing bind credentials shall continue working as normal.