Use DescribeDBInstancePages to list all RDS
We use a paginated endpoint to retrieve the RDS instances, but that
only list the first 100 RDS instances. Because that, when reseting the
master password, if there are more than 100 instances,
we are only reseting some of them.

In this commit we use the Page by Page endpoint
DescribeDBInstancePages to retrieve all the instances.