Try to drop username generated with MD5
We previously changed how we generate usernames and passwords (using SHA256 instead of MD5). But this change will cause any unbind for an existing binding to fail as the unbind tries to delete a different username (generated with SHA256 and not with MD5).

To fix this we will try to drop both the old username and the new username.