CheckAndRotateCredentials: don't defer connection closing.
The defer here is inside a for loop and therefore doesn't get triggered
until all iterations of the loop have completed and the function
returns. This means that the SQL connections to the database instances
are held open until all the instances have been checked. As the number
of instances increases, this results in a large number of connections
being held open simultaneously.

Instead, we can close the connection immediately, as there's no need to
keep it open once the password has been checked. We can also call it
unconditionally as sql.DB.Close is idempotent, and sqlengine.Close() is
a no-op if the db instance is nill.