cron: only listen for signals when started.
This was unconditionally starting the goroutine to listen for signals to
stop the cron process, which is unnecessary. Update things so that this
only happens whent he cron process is started.

This also extracts a separate function to setup the cron process, which
avoids the need for a global variable.