Add basic caching for ListTagsForResource calls
We are experiencing rate limiting throttling in our RDS brokers,
very likely due the amount of calls to the ListTagsForResource
operation. This operation is called per intance when listing the
all instances and filtering per tag to learn the broker ID.

Currently we import the same module in the
paas-rds-metrics-collector[1] project, which periodically lists
all the instances and queries the tags to discover new instances.

In our current rds-broker design these tags do not change during
all the live of the service, so it shall be fine to cache them.

We add a small private method that would cache the list of retrieved
tags indexed by ARN with no ttl.

Regarding the question: Are tags added atomically during creation?

The CreateDBInstance receives the tags on creation, but
we were concern if setting the instance Tags on creation would
not be atomic and the instance would be listed by Describe before
ListTagsForResource was returning it.

We testing this manually by querying constantly the tags while
creating an instance. All the tags appeared immediatelly in the
first response as soon as CreateDBInstance finished, so we consider safe to cache it.

[1] https://github.com/alphagov/paas-rds-metric-collector/