Start CheckAndRotateCredentials separately from HTTP server
We want to be able to move where the CheckAndRotateCredentials is
called, so we need to separate this call from the code that starts the
HTTP server.

I've also added a comment explaining why we don't use ListenAndServe
because it wasn't clear to me, and I initially refactored this before
hitting issues in the tests, and then finding the commit that introduced
this - 4a63bbe.