Move housekeeping tasks (cron) into main process.
There's no real need for these to run in a separate process. Putting
them in the main process makes the configuration and deployment easier.

This introduces a new `run_housekeeping` config variable to control
whether these run or not - they should only run on one instance.

Pulling this into the main process will make it easier to run the master
password check/rotation process on a single instance.