Ensure properties are not changed during unrelated operations (e.g. master credentials rotation).
When converting from rds.DBInstance to our own VO awsrds.DBInstance,
also include the boolean and int values like:

 - AutoMinorVersionUpgrade
 - BackupRetentionPeriod
 - CopyTagsToSnapshot
 - MultiAZ
 - PubliclyAccessible
 - StorageEncrypted

This would ensure that the operation of checking and reseting the
master password won't wipe these values during the conversion
of the objects.

Without this there is a risk of reseting these values during
the password reset, as we were convering:
rds.DBInstance => awsrds.DBInstance => rds.ModifyDBInstanceInput:

 - AutoMinorVersionUpgrade
 - CopyTagsToSnapshot
 - MultiAZ
 - PubliclyAccessible

The value BackupRetentionPeriod was not being reseted because
the function we have to convert to  the rds.ModifyDBInstanceInput
ignores this value if it is 0.

The StorageEncrypted cannot be changed.

Pending: add test