Allow force refresh of cached tags in Describe*()
In some cases we want to be able to skip the cache of tags
when calling the Describe*() methods (i.e. checking the next
change in the LastOperation() handler).

We decided to extend the interface to allow pass a list of optional
options using variant arguments for the Describe*() methods, and add
one option: `DescribeRefreshCacheOption`.

Using variant arguments allow us to add new options to the interface
without having to change the consumers.

We discarded the option of adding a InvalidateCache(ID) methods
because the current cache requires to learn the ARN of the db
instance, which would require another api call.