LastOperation refreshes tag cache of instance
The LastOperation broker endpoint relies on the reading and updating
the tags of the DB instance to implement a short of state machine
to execute the requires steps to restore an instance from snapshot[1]

In [2] we added some caching logic to avoid querying the instance
tags in every Describe() operation, but that broke the LastOperation
feature.

We need LastOperation to bypass and refresh the cache of tags every
time the awsrds.Describe() is called.

[1] https://github.com/alphagov/paas-rds-broker/pull/35
[2] https://github.com/alphagov/paas-rds-broker/pull/83