CI test helpers can create snapshots
Add methods to create arbitrary snapshots from any instance appending
an timestamp to the name. This will be used to test the restore
from snapshot.

We make the specific methods for the final snapshot more generic by
passing any snapshot ID,  and add a helper to get the final snapshot
name.

In the config of the CI tests, we add the flag `copy_tags_to_snapshot`
so the new snapshots contain always the right tags.