Integration test for restore_from_latest_snapshot_of
We test the restore for snapshot feature by:

 1. Creating an instance
 2. Creating an snapshot
 3. In backrgound deleting the first instance
 4. Create a second instance from background
 5. Delete the snapshot and the second instance.

To help managing the provisioning, waiting for it and ensure that
the cleanup happens and it happens only once, we add a small
helper ProvisionManager.

ProvisionManager performs the provisioning.
It gets a "Provisioner" function that should do the creation and
returns:
 - a function that would synchronously wait for the thing to finish
 - a function that would clean up whatever is created

The caller can optionally optionally wait for the creation to finish
can call the cleanup at any moment several times if needed.