Run 16 nodes in the integration tests
The tests are really slow but independent. There are currently 11 of
them.

Try to run all of them in parallel.

We add also add a random sleep of up of 30 seconds to ensure that
not all the workers start at the same time.

Note: Apparently Before/AfterSuite would be executed once in each node.