Use SynchronizedBeforeSuite in integration test
In the integration test we need to create some shared resources
before testing:
 - Build the binary
 - Create subnet group
 - Create security groups

We also want to run all the tests in parallel to speed up these
tests, as creating RDS instances is slow. If possible we want
to run all in parallel.

But using the normal AfterSuite/BeforeSuite methods for the creation
of the shared resources would run that in each node, not reusing
the resources.

To run a unique initialisation function in only one node we must
user SynchronizedBeforeSuite[1].

We refactor the suite for that, with two functions: one to
create the shared resources that only runs on the 1st node. Other
to read the shared data from the 1st functions in all the other
nodes (the nodes do not share memory).

We only need to share the path of the binary of rds-broker and
the generated configuration.

[1] https://godoc.org/github.com/onsi/ginkgo#SynchronizedBeforeSuite