Fix integration config to ensure unique plan IDs
In the integration tests we were using a example configuration file
which had the same `id` for the plans in MySQL and Postgres.

But how the broker is implemented, these IDs must be unique, otherwise
only the first ones would be taken into account.

In our case it resulted on the integration test not testing for mysql,
only twice for postgres.