Remove the DBInstanceDetails abstraction
The abstraction is unnecessary, and increases the risk of bugs being
introduced. This was realised recently with an incident caused by RDS
instances that were meant to be HA not being HA.

Instead we're using the types from the `rds` package directly.

Notable things
==============

Separating `GetResourceTags`
----------------------------

The old `DBInstanceDetails` type which we were passing around had `Tags`
on it, but most of the `rds.*` types do not. We considered having the
methods which returned `DBInstanceDetails` also return `Tags` as a
separate return value, but in the end we decided to extract the getting
of Tags into a separate method.

Hector voiced a concern that separating the call to get tags from the
other calls may increase the chances of race conditions where (for
example) the resource exists when Describe is called, but is deleted by
the time GetResourceTags is called.

Removing `allowMajorVersionUpgrade`
-----------------------------------

There used to be a function called `allowMajorVersionUpgrade`, which was
supposed to check if we were trying to do a version upgrade and set the
value to true if so. The function was a bit broken (it was comparing
version numbers alphabetically instead of numerically).

It seems like the intent was to always allow major version updates if
requested, so we can just default it to true and the behaviour will be
the same.

Avoiding setting DBSubnetGroupName if it's the same
---------------------------------------------------

The AWS API throws an error if you set the `DBSubnetGroupName` to the
same value as it had before - it wants you to leave it as null if you're
not planning on setting it.

We've added some code to explicitly set it to null if it's the same.

Removing check that we're not modifying the Engine
--------------------------------------------------

There used to be a check in Modify that we weren't trying to change the
database engine (e.g. from postgres to mysql) so the user would get a
nice error in that case.

Now that we're using the types from the AWS client it's not possible to
ask Modify to change the Engine (which makes sense), so we can't do the
check here.

Since the error is unlikely to happen in the wild (we'd have to change
the engine on an existing service plan), we think it's ok not to handle
it.

Removal of many tests which are now not needed
----------------------------------------------

There were several tests of the form:

```
Context("when some field is set", func() {
  BeforeEach(func() {
    ...
  })

  It("does not return error", func() {
    err := rdsDBInstance.Create(dbInstanceIdentifier, dbInstanceDetails)
    Expect(err).ToNot(HaveOccurred())
  })
})
```

which were mostly exercising the copying too and from the type we've
removed. Since these tests don't provide a lot of value they've been
removed.

Some wrapping `Contexts` have also been rolled into the tests where it
makes sense to do so.

Moving `AddTagsToResource` onto `RDSInstance`
---------------------------------------------

Previously `AddTagsToResource` lived in `utils` - it's moved onto the
`RDSInstance` which makes it a bit more convenient to call.

Switching to counterfeiter generated fakes
------------------------------------------

Since we were changing the interface of DBInstance we would have needed
to upgrade all the pre-existing fakes by hand. It's clearer and more
consistent with the rest of the codebase to use counterfeiter to
generate fakes instead of hand-coding them.

Using pointers instead of value types for configured `RDSProperties`
--------------------------------------------------------------------

This is more convenient because the `rds.*` types take pointers, so it
saves us the effort of converting all the time. Also since the AWS API
distinguishes between `null` and empty string it makes sense for our
config to have the same distinction.

Follow up work
==============

Use counterfeiter for the sqlengine and sqlprovider interfaces.

Co-authored-by: Hector Rivas Gandara <hector.rivasgandara@digital.cabinet-office.gov.uk>