Replace Travis with CircleCI
CircleCI:

* makes it much easier to run our unit tests against multiple Postgres versions
* has a configuration language is also much easier to read and understand
* seems to queue and run builds much quicker than Travis

However, CircleCI does try to be 'helpful' and detects when you are
trying to run a Postgres service (even ones in Docker images they don't
own) and will configure it to trust all passwords. This breaks tests of
login failures. To work around this we had to add:

```yaml
POSTGRES_INITDB_ARGS: "--auth-host=md5"
```

The Docker image is based on the official one[1] and that's the
documented way of overriding stuff via `initdb`[1].

[1] https://hub.docker.com/_/postgres/