Allow unit testing of multiple Postgres versions
The broker needs to support Postgres 9.5 and 10.5.

The MySQL Docker container takes a while to start, so we only want to
start it once at the start of the unit tests, and tear it down at the
end.

We need to run the same tests against multiple Postgres versions, so we
run all unit tests first against Postgres 9.5, then just repeat the
`sqlengine` package against Postgres 10.5.

I've observed some test failures in both MySQL _and_ Postgres, caused by
the servers not being ready by the time the tests start running. For now
I have fixed this with `sleep`s, but perhaps a better solution could be
found.