Implement reboot as a update operation
We want to enable the tenants to trigger a instance reboot if they
desire. We implement the option to reboot the instance by accepting
custom update parameters.

The reboot must be used alone, and any plan change is passed it will
make the update fail. We do not check service changes because it is
an immutable value[1]

We also allow to force the failover of a HA RDS instance.

[1] https://github.com/openservicebrokerapi/servicebroker/blob/v2.12/spec.md#updating-a-service-instance