Allow self-service provisioning of supported Postgres extensions
Users have requested that Postgres extensions be enabled on existing databases
in the past, which has required manual intervention. In this commit, users can
self-serve supported extensions for Postgresql (specified in configuration, per
plan) when creating the service e.g.

    cf create-service postgres tiny-unencrypted-9.5 \
      -c '{"enabled_extensions": ["postgis", "pg_stat_statements"]}'

Some extensions require shared preload libraries to be enabled in the AWS RDS
parameter group. To support this, parameter group names are supplied in the
configuration file, and decoded to determine which special case extensions they
support. E.g.

    rdsbroker-postgres95-envname-pg-stat-statements

The parameter group names are used in conjuction with the requested extensions
to determine which parameter group the RDS instance should be created in.