Create RDS parameter group names on demand using an internal naming convention
Rather than have parameter groups passed in with structured, parseable names,
allow the RDS broker to create its own parameter groups. This change allows the
parameter group names to be more flexible, because they're not required to be
parsed. It also reduces the coupling between paas-cf and paas-rds-broker,
because they no longer need to agree on a naming convention.