Add `engine_family` property to plan configuration
RDS engine family strings aren't always the engine and the version
concatenated, for example in the case of postgres 10.5. To keep things simple,
a new `engine_family` property has been added to plan configuration. Its value
should be the correct family name for that engine and version, as reported by
`aws rds describe-db-engine-versions`.

Documentation and config samples are updated to include the new property.