Fix handling of SkipFinalSnapshot tag.
The tests for the Provision handling of this flag were broken because
they weren't passing in the correct JSON, and were only accidentally
passing because they changed the plan in use. This fixes the tests to be
consistent with each other other than the variable under test.

The Update handling of this flag was also incorrect in that it would
reset the value back to the plan default value unless the user passed in
the value on every update call. This updates the behaviour so that the
tag is only changes on update if the user explicitly passes in a new
value. This also adds some tests for this case.