Error for extra fields in user-provided JSON.
Prior to this change, any unrecognised params in the user-provided JSON
would be ignored. This means that it was easy to make a typo, and think
that a change has been applied when in reality it's been ignored.

Returning an error for any extra fields in the provided JSON will make
it easier to spot the problem.