Add a delay to pollForOperationCompletion
Sometimes the AWS API takes a few seconds to start modifying an
instance. If we poll before the operation has started, it will return
immediately, and then later operations will fail when attempting to do
further operations on the instance.

Adding a delay before starting to poll should prevent this. This is also
more inline with how CloudFoundry will interact with the broker as it
doesn't call the lastoperation endpoint immediately after an update has
completed, it instead waits a little bit.