Check compatibility of plan change with parameters
We currently just check that the user hasn't asked to reboot the
instance at the same time as changing the plan, because that combination
is not valid.

We'd also like to specify that certain other combinations of parameters
are not valid when changing the plan, for example enabling or disabling
extensions.

This commit refactors the existing check to make it a bit easier to add
new ones.