Update parameter group when updating plans
Prior to this commit we'd only update the parameter group if the
extensions were changing. This broke updates when the plan was changing
from an older engine version to a newer version - the parameter group
needs to change here too.

In previous commits we've validated the input to ensure there is never a
situation where the user is both changing the plan and the extensions,
which could otherwise cause a confusing situation trying to work out
whether the database needed restarting to support any changes to preload
extensions.