Explicitly disable upgrades from postgres 9
These are currently broken due to an issue with the postgis plugin
(which we enable by default). Since the failure mode is quite bad (the
upgrade looks like it has succeeded, but it actually hasn't worked) we
think the best bet is just to explicitly disable this upgrade and tell
people to contact support.