Keep patch level modifying similar EngineVersion
We are not prescribing the patch level in the RDS service catalog
plans in order to make use of the auto minor version upgrade feature.

This means that AWS chooses the patch level when creating or modifying
an instance. If the patch-level chosen for the modification is for some
reason lower than the patch-level of the current instance an update will
fail.

This change preserves the patch-level of the former instance. if the
major and minor versions are the same. It also adds tests.