Stop unnecessary connection when updating Postgres
A tenant has completely filled up their RDS Postgres instance and wants to upgrade to a larger plan. The broker should be able to handle this fine, but it doesn't in practice. In practice the broker is failing to process the update, because it errors when it can't connect to the Postgres instance.

This commit removes a situation where the RDS Broker connects to Postgres instances even when there is no work to be done. I'm not certain that this is the only reason changing plan is failing, but it's at least one of the reasons.