Enable broker to pick a compatible upgrade target
When performing a version upgrade, AWS RDS has specific lines of
compatibility; version 9.5.x may be able to upgrade to 10.14, but
version 9.5.(x-1) might not. Historically we've always used the
version written in to the config as the version we want AWS RDS to
upgrade the database to. If that version isn't EXACT (e.g. Postgres
10), AWS RDS will assume we want to upgrade to the most recent minor
version (e.g. Postgres 10.14).

Given AWS RDS' specific compatibility rules, specifying an
incomplete target version can result in being unable to upgrade the
database, and the tenant being powerless to do anything about it.

Instead, this commit adds some intelligence to the broker. When an
upgrade is initiated, it will look up the most recent valid upgrade
target for the version of the database being upgraded to.

To explain, consider a plan change from `small-ha-9.5` to
`small-ha-11`.

The broker will now look up the exact version currently being run
(e.g. 9.5.23), and then look at what versions RDS says are valid
upgrade targets (e.g. 10.14, 10.15, 11.0, 11.1, 12.2). It will then
filter those targets by the version specified in the new service
plan (results in 11.0 and 11.1). Finally, it will pick the most
recent version, according to semantic versioning rules: 11.1.

Thus version 11.1 will be sent to AWS RDS as the engine version to
upgrade to.