Prevent upgrades from Postgres 9.5 to 11 or higher
Some of our Postgres 11 and higher plans use instance types that are not
supported for Postgres 9.5, This is a problem for us, because AWS RDS does a
database upgrade with a simultaneous instance type change in two steps: 1.
Change the instance type
2. Change the engine version

When a tenant tries to upgrade from Postgres 9.5 to Postgres 11 or higher, and
the target plan uses an instance type not supported for Postgres 9.5, AWS RDS
will give us an error, because of the steps above.

We could change the broker to do a version and instance class change in two
steps, in the opposite order to AWS RDS, but this would add extra complexity to
the broker. This isn't a sensible plan when there Postgres 9.5 EOL is
impending, and many databases will need to be upgraded.

Instead, we instruct tenants to upgrade to Postgres 10 first. Our version 10
and version 9.5 plans use the same instance types, and Postgres 10 is supported
on all of the instance types used in the Postgres 11 and above. This a neat
workaround, but does put extra labour on the tenant.