Wait for databases to be cleaned up in integration tests
It's not possible to delete resources like security groups and
parameter groups when there are >0 databases depending on them. Test
runs would fail in the past because the databases hadn't been fully
deleted before those things were deleted.

We wait for the databases to be deleted so that doesn't happen.