Build Bosh release via Make
Creates a Make task that can be invoked to build a Bosh release for the RDS
broker. It has sensible defaults, and they can be overriden for use in a CI
pipeline.

The one thing to note is that it copies the contents of the repo, excluding the
release directory, in to the source path for the release. It does not use a
symlink, because that would result in a loop.