Raises the number of integration test nodes for greater test parallelisation
Ginkgo runs a number of nodes (processes) and divvies up the tests to be run
among them. The tests on each node run in parallel. All of our database
integration tests are independent from one another, so they're ideal for
parallelisation.

Another nice benefit is that they should finish a little quicker.

Also updates the slow spec threshold to be a little more realistic about what a
slow spec is.

Also makes integration tests stop on the first failure, in order to reduce
cycle times when a test fails.