Run only unit tests at appropriate time in CI workflow
Previously, we were running `make unit` and THEN the SQL engine
integration tests as part of the CI pipeline.

`make unit` internally runs the unit tests AND the SQL engine tests,
so the end result was the SQL engine tests being run twice.

The correct Make target is `test_unit`.