Reduce parallelisation of broker integration test
Ginkgo parallelises tests by spinning up a number of nodes
(processes), and farming out the running of individual test cases to
them. The number of nodes is configurable at the command line.

The AWS RDS API has a rate limit, which isn't documented. Our
Technical Account Manager says it isn't documented because it's
internal and not shareable.

With 18 nodes, we had 18 tests running simultaneously. That caused
us to regularly hit the rate limit. This commit reduces the number
of nodes, so that we don't hit it.