add check_pg_extensions script
this is the script that was used to test insertion of all pg extensions
for the postgres13 offering. if needed, someone might like to convert
this into a proper integration-style test. the problems with that being
that such a test would need to be provided with:

 - the full list of extensions we're considering allowing for an
   offering, but this information tends to live in the paas-cf
   repo.
 - the ordered dependencies of each extension, which we don't
   currently record anywhere programmatically.

it might be slightly neater to redesign this to create a new database
for each extension and specify all dependencies together in a single
enable_extensions at creation time. this would run much faster too.

parallelization would also be a possibility if you weren't afraid of
rds rate limits or number of databases limits present in a dev
account.