Exclude temp tables from read-only access trigger
The trigger make_readable is executed on a number of CREATE commands and
grants SELECT on all tables and sequences as well as USAGE on all
schemata.
If a table is created within a stored procedure, this trigger does not
return and keeps an AccessExlusiveLock on all tables. This is effectively
globally locking the entire database for the runtime of the stored
procedure, which may be substantial (The billing update_resources
function runs for 2h+ in London).

This modification evaluates the schema_name of the
pg_event_trigger_ddl_commands() to prevent the trigger from being
executed if a temporary table is created. As we are using temporary
tables in the stored procedures this will prevent locking the database
for their runtime.

We do not envision this change to impact the read-only functionality for
tenants and read-only users should not need to access temporary tables -
TBD.

pair: @whpearson @schmie